Refactors validators
