+++ src/app/shared/validators/RegisteredEmailValidator.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AbstractControl, AsyncValidator, ValidationErrors } from '@angular/forms';
 import { Observable } from 'rxjs';
-import { UserService } from './user.service';
+import { UserService } from '../../core/user.service';
 
 @Injectable({ providedIn:'root' })
 export class RegisteredEmailValidator implements AsyncValidator {
